Wall Jump Functional
The Super Metroid-inspired wall jump now works as intended. Coding is finished for today; tomorrow will be whatever I can finish in my remaining 4 hours, but the goal will be to get the missile, parry and SFX all fully working. DO NOT FORGET THE O AND P COMMANDS WITH UI!